Implemented basic STATUS_HILITES support!
This rips off a bunch of core logic from the TTY implementation, but
reworks the logic to put the ANSI escape codes inside a buffer with the
working text instead of writing straight out to a terminal.

As part of this, I've enhanced the (very basic) ANSI escape code support
in font-bdf to handle multiple formatting directives.

This still uses the same WIN_STATUS implementation as before, but alters
the rendering to use the new window port implementation.  The next step
would be to look at reworking the rendering to take advantage of the
window port to relocate status items to make better use of screen
real-estate.