Commit the darn NDS core Makefile
Thanks to this being included in the gitignore by default, this file
never got included when I created this new vanilla port.

Upshot is no one could build it but me because I happened to have that
file on my local filesystem.

Until, that is, I moved to a new machine and started again from a fresh
checkout.

Yeah, oops.