Major tile rendering refactoring
This work is laying the groundwork for changing the way the graphics
tile loader works.

Here we pull out all the tile loading/rendering code into a separate set
of structures with some bits of low-level polymorphism to keep the
outside interface nice and simple.

This will let me mess around with the internal machinery without having
to touch any of the dependent code.