Dynamically load graphic tiles from file
In the old, simpler implementation, NHDS loaded the entire tilemap file
into memory and then pulled tiles from that cached data.  The upside is
this was simple and also pretty quick.  The downside is that large
tilesets simply couldn't be loaded as we'd run out of memory.

After the previous refactoring, this has been completely reworked so
that tile data is loaded from the file directly as its needed.  That
means big tilesets work well, and we have a much lower overall memory
footprint which I hope will help with stability.

This first implementation is pretty simple, and not actually super
efficient in that we load whole rows of tiles at a time when we need a
tile from the file.  Now, the good news is for a lot of tiles, they're
logically grouped, so a little caching can help.  But it's honestly not
the fastest implementation.

However, I'm going to stick with this for now and see how it goes.

If it turns out performance is unacceptable, the next step is to build
an on-disk cache of the tile data that's reorganized so it can be very
quickly read from disk dynamically.  But premature optimization and all
that...

And while I was at it, I also fixed up the tile caching logic a bit, as
it turns out with graphics tiles we were evicting way more frequently
that needed.