Hijack the NH patchlevel field...
I needed some way to distinguish NHDS version releases, so we're using
the patchlevel for now.  It's not ideal, but the alternative is some
other version distinguisher and I don't want to introduce that if I can
avoid it.

For now, we're still using X.Y.Z in the data file directory name, as
savefiles and so forth should work across every version in a patchlevel,
but if it becomes necessary to distinguish them, I can add support for
using directories with the patch level included as well.