Don't allow menu dismiss if no item selected
Because NHDS uses a slightly odd menu input scheme (X/Y for
select/unselect, A/B for accept/dismiss), it's incredibly easy to
accidentally hit A when you really meant to hit X, with the result being
pretty darn confusing. This is doubly true because the typical idiom on
the DS is or A to be "OK" and B to be "Cancel", so for a single item
select menu (e.g. a yes/no question), it's easy to highlight an item and
press A when you really meant to press X (I'm tempted to make these
synonym actions for single item select menus, but not yet).

So this change simply stops users from doing something they probably
never meant to do: accept a menu without actually selecting an item.

I'm also entertaining putting up a prompt message telling the user what
went wrong if they attempt this, but I gotta figure out how to make that
play nicely in the character picker workflow before the statuslines are
initialized...