Further improve single menu item select behaviour
As mentioned in the previous commit, I wanted to try supporting
highlighting an item and pressing A to select items in a single item
selection picker.

I liked this a lot *but* commands that use the yn_function are tough
because, without core changes, there's no way to know, within the
nds_yn_function implementation, if a command can take a count or not,
since that information is part of the inventory item's class, and that
isn't available at that point.  The only information we have is the
question being asked, which includes a set of choices that we parse out
in order to display our menu.

So, we get ugly.

This commit includes a core change to invent.c so that getobj includes,
in the prompt string, and string indicator if the command can take a
count. To be honest, this is probably just useful information for the
user in the prompt, anyway, since folks might not be aware that, say,
'drop' can take a count. But in the port, I can use this indicator to
figure out if a count is allowed or not, and alter the menu behaviour
accordingly.

The end result is better menu behaviour in single- vs multi-pick
scenarios, and hopefully a less confusing user interaction model.

Assuming I didn't break anything, of course.