Fixed another memory stomper!
It's official, I'm a wizard.

For a very long time now I'd noticed the debug output window would
occasionally get flooded with '@' characters.  But I'd just given up
trying to hunt down the issue since it wasn't catastrophic, just
annoying.

Well, I finally decided to look into the bit pattern of that data, and
given an '@' sign is ASCII value 64, the bit pattern is a repeated
sequence of '0100' bits, which is what you'd expect from out-of-bound
graphics data.

Eventually I was able to consistently reproduce the issue by (my
character) eating until satiated, and I noticed in that moment, the
status window buffer increases by one line while the message window
moves down.

Welp, turns out, when that happens, I wasn't shrinking the vertical size
of the PPM to compensate.  The result is that the image was getting
drawn past the end of that BG layer memory and into the next layer
memory, which is the character buffer for the debug window.

Hence this fix.

Here, we check if the height of the message window doesn't match the
height of the allocated image.  If they differ, the image is free'd and
then reallocated.

And while I was in here I removed some old code I'd left behind for
testing purposes... and no, I absolutely do not remember what I was
doing at the time.  T'would be ironic if I was investigating this bug.