Make direction prompting less error prone
The direction prompting code just uses nds_get_input, which is the same
entrypoint for getting general input in the game.  That means any
keypress is valid, and the press is then passed through the keymap,
which determines the character that's returned.

This is good because it ensures direction selection supports whatever
direction keys were bound.

It's bad because you can provide invalid input, which then causes
'cmdassist' mode to kick in (by default), which dumps some help to the
screen that's not actually helpful.

This is particularly frustrating because if you want to cancel whatever
operation has been triggered (throw, zap, etc), the instinct is to press
the B button, but that just returns whatever B has been bound to (which
by default is the 'kick' command), and so you get a weird error instead
of expected behaviour.

So this code has been cleaned up in two ways:

1. nds_yn_function will re-prompt the user until they provide a valid
   input, by confirming the user pressed a direction key.
2. Pressing B properly cancels the input.  This is done by including a
   new options flag (which right now only has one value) for
   nds_get_input that changes the behaviour so, when B is pressed,
   instead of returning the bound key, the first quitchar is returned
   first.  That quitchar is then returned in nds_yn_function.

The upshot is, now if you start a command that takes a direction input,
you either have to provide a valid input, or you can cancel with B, and
everything Just Works.