Enable clicklook (and fix tap-and-hold behaviour)
By enabling clicklook, and along with the additional changes in this
changeset, the user can now tap-and-hold on the map to identify items
instead of using the 'whatis' command.

Note, this does revert a previous change I'd made where CLICK2 was being
remapped to CLICK1 in the normal game loop.  The commit message
associated with this change mentioned making that modification to better
support compassmode, and I'll be honest, I don't remember why.  If I do
uncover why, I can always reintroduce that remapping by conditioning it
on compassmode being enabled.