Fix inventory window getting clipped
The inventory window is persistent, and if its contents change, the
calculated width of the buffer to put on the screen can also change.
When this happens, we need to reallocate a wider image to draw into.
This change causes the code to detect if that reallocation must happen,
and then does so if needed.