Bind Start to #herecmdmenu by default
So turns out the herecmd menu is enormously useful in this port as it
makes it easy to access commands relevant to the current player location
(opening chests, eating corpses, etc).  So we're gonna bind Start to
invoke the menu by default.

Also updating the doc accordingly, but of course Lyx made a bunch of
extraneous changes so.. yeah.  Really gonna just switch the manual to
Markdown...