Enable EDIT_GETLIN and set default player name
If the personal data contains a name composed of ASCII characters, the
NetHack user name now defaults to the name entered on the device (Sorry
non-English players!  I'd love to support unicode names, but NetHack
doesn't...).

This change also enables EDIT_GETLIN which causes the input prompt to
default to previously entered values in various circumstances, which is
convenient.